requirements.txt
opencv-python==4.6.0.66
imutils==0.5.4
mediapipe==0.8.10.1
Pillow==9.2.0
numpy==1.21.6